Added culture-sensitive and invariant Parse/ToString functionality
Modified the Parse and ToString functionality to take into account
current culture settings, and added ParseInvariant and ToStringInvariant
that use an invariant culture.

Created a new public static class, SqlArrayCultureFormatter to handle
the difference in the interface of the regular types, and the SqlComplex
type.

Had to modify the ArrayParser project to take into account culture
information.